LPS-77727 Move to a proper package
